Make crypto profiles actually trade, and trade as crypto rather than NSE equity
The Crypto Trading tab had never placed a single trade. Three separate
defects, each of which alone was enough to prevent it:

1. crypto_momentum returned plain dicts where the shared entry path
   reads cand.symbol/cand.ltp, so every scan that found a candidate
   raised AttributeError. run_forever()'s catch-all swallowed it, so the
   profile logged candidates every cycle while silently crashing and
   never buying. It now returns a Candidate dataclass like every other
   strategy.

2. The buy-reason builder's fallback branch named momentum_52w_high's
   own fields (pct_from_52w_high, relative_strength_pct, volume_multiple)
   unconditionally, after the buy was sized and approved but before
   broker.buy() ran. That crashed FOUR profiles -- crypto_momentum,
   pivot_supertrend, trend_pullback and long_term_trend -- none of which
   have those fields. Built from getattr now, so a strategy contributes
   whatever it actually carries.

3. Position sizing floored to whole units (int(budget // price)), an NSE
   whole-share assumption. Crypto is divisible, so BTC-USD at ~84,000
   against an 8,000 budget sized to exactly 0 and was unbuyable outright,
   while ETH-USD sized to 2 and left a third of the budget idle.
   RiskManager takes a fractional_quantities flag (derived from the
   profile's category, like trades_24_7), quantities are floats through
   models/broker/schema, and sells use a dust epsilon so a full exit
   can't strand an un-closeable 1e-17 remainder. Equity profiles keep
   whole-share sizing and are bit-for-bit unchanged.

Removed the NSE/Nifty coupling that remained in the crypto path:

- Liquidity: min_avg_daily_turnover_inr is an NSE rupee figure, but
  crypto turnover is quoted in the pair's USD quote currency, so the
  inherited 50,000,000 demanded $50M/day -- ~83x stricter than the ~₹5cr
  ($600K) it means for equities. It rejected 16 of 24 coins as
  "illiquid", including LINK (~$32M/day) and ARB (~$43M/day). Crypto
  profiles now set min_avg_daily_turnover_usd: 5,000,000, which the
  crypto strategies prefer over the _inr field. Rejections fell 16 -> 3
  and candidates rose 7 -> 18.
- Fees: execution: is now profile-scoped (like strategy:/risk:), so
  crypto charges a percentage of notional (fee_pct: 0.1, the ~0.1% taker
  fee a real exchange charges) instead of an Indian broker's flat ₹20
  per order. Equity profiles keep flat_charges_inr: 20 exactly as before.
- The consolidation_breakout buy reason hardcoded ₹ even though the same
  strategy runs the USD-quoted Crypto Breakout profile.

Also fixes a latent correctness bug found while testing: a flat window
has zero average gain AND zero average loss, so RSI is 0/0 = NaN. Every
comparison against NaN is False, so a NaN RSI sailed straight through the
`rsi < min_rsi` entry gate -- a dead, non-moving coin was treated as
having momentum. _rsi now returns None for that case (RSI 100 on a
zero-loss uptrend is still preserved), and it is reported as its own
no_price_movement rejection rather than being mislabelled
insufficient_history in the dashboard's breakdown.

The two crypto tests that had been failing since before this change were
asserting exit reasons their own fixtures couldn't produce (flat closes
give an undefined RSI, not a low one, and the RSI exit is checked before
the trend break). Rewritten against data that isolates each condition.
Full suite now passes at 158, up from 152 passing / 2 failing.

Verified live: crypto_momentum opened 3 fractional positions
(1818.18181818 NEAR, 37558.685446 ARB, 885.73959255 UNI), each deploying
the full 8% budget, each charged 8.00 = 0.1% of notional, with no
crashes. 52w_high (26 trades) and cross_sectional (15) are unaffected.

+++ src/papertrader/portfolio/models.py
@@ -8,7 +8,10 @@ from datetime import datetime
 @dataclass
 class Position:
     symbol: str
-    quantity: int
+    # float, not int: crypto trades in fractions (0.095 BTC is an
+    # ordinary size). An equity profile still only ever puts whole
+    # numbers in here -- see RiskManager.fractional_quantities.
+    quantity: float
     avg_price: float
     entry_date: str
     highest_close_since_entry: float
@@ -34,7 +37,7 @@ class Trade:
     id: int | None
     symbol: str
     side: str  # BUY or SELL
-    quantity: int
+    quantity: float  # fractional for crypto -- see Position.quantity
     price: float
     charges: float
     reason: str
